:art: FLOW-647 render code segregated
